refactor outbox and tests
move outbox lib inside controller directly since no one else is using it
only have api tests (moved to outbox)
