+++ tests/api/utils/activities.js
@@ -1,7 +1,6 @@
 const CONFIG = require('config')
 const host = CONFIG.fullPublicHost()
 const { getRandomBytes } = require('lib/crypto')
-const { byUsername } = require('controllers/activitypub/lib/activities')
 
 const randomActivityId = (origin = host) => `${origin}/${getRandomBytes(20, 'hex')}`
 
@@ -16,6 +15,4 @@ const randomActivity = ({ externalId, emitterActorUrl, activityObject, type }) =
   }
 }
 
-const getActivitiesByUsername = username => byUsername(username)
-
-module.exports = { randomActivityId, randomActivity, getActivitiesByUsername }
+module.exports = { randomActivityId, randomActivity }
